Closing watches is a bit redundant
When the transport is closed, the OS cleans up the inotify watch
descriptors with no API to Watch objects. It is convenient enough
to react on the transport closing compared to individual Watch
instances.

Let us look later into what happens with Watch.close() when the
descriptor has been externally removed.